Consolidate schema extraction logic
